+++ PIS_PROYECTO/data/partidopolitico.json
@@ -1,1 +1,1 @@
-{"controlador.ed.listas.ListaEnlazada":{"cabecera":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":1,"nombre_partidoPolitico":"Alianza País","esloganPartido":"\"5\""},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":2,"nombre_partidoPolitico":"Creo","esloganPartido":"\"21\""},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":3,"nombre_partidoPolitico":"Alianza Union Por La Esperanza","esloganPartido":"\"1!"},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":4,"nombre_partidoPolitico":"Revolucion Ciudadana","esloganPartido":"\"6\""},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":5,"nombre_partidoPolitico":"Izquierda Democratica","esloganPartido":"\"12\""},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":6,"nombre_partidoPolitico":"Contigo Avanza","esloganPartido":"\"35\""},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":7,"nombre_partidoPolitico":"","esloganPartido":4},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":8,"nombre_partidoPolitico":"","esloganPartido":4},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":9,"nombre_partidoPolitico":"","numero_Lista":4}}}}}}}}}},"size":9}}
+{"controlador.ed.listas.ListaEnlazada":{"cabecera":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":1,"nombre_partidoPolitico":"Alianza País","esloganPartido":"\"5\""},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":2,"nombre_partidoPolitico":"Creo","esloganPartido":"\"21\""},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":3,"nombre_partidoPolitico":"Alianza Union Por La Esperanza","esloganPartido":"\"1!"},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":4,"nombre_partidoPolitico":"Revolucion Ciudadana","esloganPartido":"\"6\""},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":5,"nombre_partidoPolitico":"Izquierda Democratica","esloganPartido":"\"12\""},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":6,"nombre_partidoPolitico":"Contigo Avanza","esloganPartido":"\"35\""},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":7,"nombre_partidoPolitico":"","esloganPartido":4},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":8,"nombre_partidoPolitico":"","esloganPartido":4},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":9,"nombre_partidoPolitico":"","numero_Lista":4},"sig":{"info":{"@class":"modelo.PartidoPolitico","id_partidoPolitico":10,"nombre_partidoPolitico":"","numero_Lista":""}}}}}}}}}}},"size":10}}
